Inlining `DataSource.update(…)…`
Let's first inline both `DataSoruce.update(…)…` and `.read(…)…` then extract common patterns inside them.
